fix: :bug: Fix order of allowing users from homeserver
Previously, the Bot class tried to add the homeserver users  to the allowlist during Bot class init. This raises an error because api.async_client has not been created yet. This change sets a private attribute that is checked upon bot main, and allows the users from the same homeserver if that is the case.